Fix tag name, and pull re-used logic to trait to break inheritance issue with bulksql getting the bulkcsv interpreter tags.
